requirements.txt
altgraph==0.10.1
aniso8601==1.2.0
bdist-mpkg==0.4.4
beautifulsoup4==4.3.2
bonjour-py==0.3
click==6.6
Django==1.10.3
Flask==0.10.1
Flask-Admin==1.4.2
Flask-HTMLmin==1.2
Flask-Login==0.4.0
Flask-RESTful==0.3.5
Flask-SQLAlchemy==2.1
get==0.0.0
gunicorn==19.3.0
htmlmin==0.1.10
image==1.5.5
itsdangerous==0.24
Jinja2==2.7.3
lxml==3.4.4
macholib==1.5
MarkupSafe==0.23
matplotlib==1.1.1
mechanize==0.2.5
modulegraph==0.10.1
numpy==1.6.2
Pillow==3.4.2
post==0.0.0
public==0.0.0
py2app==0.7.1
pyobjc-core==2.3.2a0
pyobjc-framework-AddressBook==2.3.2a0
pyobjc-framework-AppleScriptKit==2.3.2a0
pyobjc-framework-AppleScriptObjC==2.3.2a0
pyobjc-framework-Automator==2.3.2a0
pyobjc-framework-CFNetwork==2.3.2a0
pyobjc-framework-Cocoa==2.3.2a0
pyobjc-framework-Collaboration==2.3.2a0
pyobjc-framework-CoreData==2.3.2a0
pyobjc-framework-CoreLocation==2.3.2a0
pyobjc-framework-CoreText==2.3.2a0
pyobjc-framework-DictionaryServices==2.3.2a0
pyobjc-framework-ExceptionHandling==2.3.2a0
pyobjc-framework-FSEvents==2.3.2a0
pyobjc-framework-InputMethodKit==2.3.2a0
pyobjc-framework-InstallerPlugins==2.3.2a0
pyobjc-framework-InstantMessage==2.3.2a0
pyobjc-framework-InterfaceBuilderKit==2.3.2a0
pyobjc-framework-LatentSemanticMapping==2.3.2a0
pyobjc-framework-LaunchServices==2.3.2a0
pyobjc-framework-Message==2.3.2a0
pyobjc-framework-OpenDirectory==2.3.2a0
pyobjc-framework-PreferencePanes==2.3.2a0
pyobjc-framework-PubSub==2.3.2a0
pyobjc-framework-QTKit==2.3.2a0
pyobjc-framework-Quartz==2.3.2a0
pyobjc-framework-ScreenSaver==2.3.2a0
pyobjc-framework-ScriptingBridge==2.3.2a0
pyobjc-framework-SearchKit==2.3.2a0
pyobjc-framework-ServiceManagement==2.3.2a0
pyobjc-framework-SyncServices==2.3.2a0
pyobjc-framework-SystemConfiguration==2.3.2a0
pyobjc-framework-WebKit==2.3.2a0
pyobjc-framework-XgridFoundation==2.3.2a0
pyOpenSSL==0.13
PyRSS2Gen==1.0.0
python-dateutil==1.5
pytz===2012d
query-string==0.0.0
request==0.0.0
requests==2.7.0
scipy==0.11.0
setupfiles==0.0.0
six==1.10.0
SQLAlchemy==1.1.3
Twisted==12.2.0
vboxapi==1.0
virtualenv==15.0.3
Werkzeug==0.10.4
WTForms==2.1
xattr==0.6.4
zope.interface==3.8.0
